Update transforms for table sections https://bugs.webkit.org/show_bug.cgi?id=245016
Reviewed by Simon Fraser.

The computed anchor point for transforms is based on the resolved transform origin value. This transform origin value depends on the final geometry of the associated renderer.
This patch updates it by calling updateLayerTransform() when the table section geometry is set.
(Without this fix, the anchor point computed as if the box was zero sized -e.g. center behaves as top/left.)

* LayoutTests/TestExpectations:
* Source/WebCore/rendering/RenderTableSection.cpp:
(WebCore::RenderTableSection::layoutRows):

Canonical link: https://commits.webkit.org/254351@main