[macOS] Memory leaks because of reference cycle created by -[WebRevealHighlight setClearTextIndicator:] https://bugs.webkit.org/show_bug.cgi?id=245003 rdar://99411435
Reviewed by Tim Horton.

* Source/WebCore/editing/cocoa/DictionaryLookup.mm:
(-[WebRevealHighlight revealContext:stopHighlightingItem:]): Use std::exchange instead
of just WTFMove to get and clear the "clear text indicator" function. Not needed to fix
the bug, but a more correct idiom, since WTFMove is not guaranteed to clear out the object
being moved from.

(WebCore::showPopupOrCreateAnimationController): Moved the null check of clearTextIndicator
outside the block. This is not needed to fix the bug, but the function can't become null
later, so it can be checked first, and when there is no function it is correct and safe to
not call setClearTextIndicator: on the newly created WebRevealHighlight object. Removed the
unnecessary capture of webHighlight. This is the bug fix. Capturing this created a reference
cycle that wouldn't be broken until the -[WebRevealHighlight revealContext:stopHighlightingItem:]
method is called, and was unnecessary.

Canonical link: https://commits.webkit.org/254359@main