[CMake] Share utilities build after 254382@main https://bugs.webkit.org/show_bug.cgi?id=245089
Reviewed by Michael Catanzaro.

In 254382@main all of the `TestWebKitAPI::Util` code in `Utilities.h`
was consolidated for all non-Cocoa ports. When updating the build the
ports specific implementation was just replaced with `Utilities.cpp`.
Instead of specifying it per port the `Utilities.cpp` is added to the
sources. For Cocoa ports the implementation in there is guarded with
`!PLATFORM(COCOA)` so it is safe to add it.

* Tools/TestWebKitAPI/CMakeLists.txt:
* Tools/TestWebKitAPI/PlatformGTK.cmake:
* Tools/TestWebKitAPI/PlatformJSCOnly.cmake:
* Tools/TestWebKitAPI/PlatformPlayStation.cmake:
* Tools/TestWebKitAPI/PlatformWPE.cmake:
* Tools/TestWebKitAPI/PlatformWin.cmake:

Canonical link: https://commits.webkit.org/254412@main