AX ITM: Some non-empty data tables are exposed as 0 columns, 0 rows. https://bugs.webkit.org/show_bug.cgi?id=244954 rdar://98886794
Reviewed by Chris Fleizach.

The cause of the problem was that AccessibilityTable::computeIsTableExposableThroughAccessibility was returning false if hasARIARole() return true. That didn't account for the case where a <table> could have the attribute role="grid" or role="table". This patch renames hasARIARole to hasNonTableARIARole and fixes its implementation.
In addition, enhanced AXObjectCache::handleRoleChanged to better account for a wider range of possible table and row role changes. This makes the update of the AXObjectCache and the AXIsolatedTree more robust to "arbitrary" roles for tables and table rows.

Canonical link: https://commits.webkit.org/254418@main