'animation-foo' declarations in @keyframes should be parse error https://bugs.webkit.org/show_bug.cgi?id=243537 <rdar://98503617>
Reviewed by Antti Koivisto.

The family of animation-* properties, including the animation shorthand, are
invalid in a @keyframes rule. We import the relevant WPT test which used to
fail but now passes as of this change. Some other existing tests need rebasing
since we changed the serialization to not output the vendor prefix, and also
because they are passing thanks to this change.

* LayoutTests/http/tests/css/shared-stylesheet-mutation-preconstruct-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/parsing/keyframes-allowed-properties-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/parsing/keyframes-allowed-properties.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/js/001-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/CSSKeyframesRule-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/rule-restrictions-expected.txt:
* Source/WebCore/css/CSSKeyframesRule.cpp:
(WebCore::CSSKeyframesRule::cssText const):
* Source/WebCore/css/parser/CSSPropertyParser.cpp:
(WebCore::CSSPropertyParser::parseValue):
(WebCore::CSSPropertyParser::parseKeyframeDescriptor):
* Source/WebCore/css/parser/CSSPropertyParser.h:

Canonical link: https://commits.webkit.org/254468@main