REGRESSION(250548@main): Early WebProcess exit on 32-bit platforms since memory limit is zero https://bugs.webkit.org/show_bug.cgi?id=245222
Reviewed by Ben Nham and Chris Dumez.

In 250548@main, the memory limit in MemoryPressureHandler is set to a
least 7 GB for active process.
This value is stored in a size_t variable, but this is limited to 4 GB
on 32-bit platforms.

The memory limit is truncated down to 0, which triggers a WebProcess
exit after the first MemoryPressure check.

* Source/WTF/wtf/MemoryPressureHandler.cpp:
(WTF::thresholdForMemoryKillOfActiveProcess):
(WTF::thresholdForMemoryKillOfInactiveProcess):
Also use CPU(ADDRESS64)

Canonical link: https://commits.webkit.org/254533@main