Merge 254548@main - Holepunch missing after video source change https://bugs.webkit.org/show_bug.cgi?id=245135
Reviewed by Žan Doberšek.

When adopting the new state in the scene, invalidate the proxies that are not used anymore
before activating the new ones.

This fixes a situation where a layer's proxy is replaced with a new one. For this to work
properly, the old proxy needs to be invalidated before the new one is activated, or the
invalidation will clear the contentLayer that was set by the new proxy to the associated
TextureMapperLayer.

* Source/WebKit/Shared/CoordinatedGraphics/CoordinatedGraphicsScene.cpp:
(WebKit::CoordinatedGraphicsScene::updateSceneState):

Canonical link: https://commits.webkit.org/254548@main

(cherry picked from commit f994a1b3e3dc9b3d2eae06e7464b0fd2fd70ce2c)