[GStreamer][MSE] Avoid seeking in READY or NULL pipeline state https://bugs.webkit.org/show_bug.cgi?id=245358
Reviewed by NOBODY (OOPS!).

We should put the seek request on pending when the GStreamer pipeline is in a state below PAUSED state.

Because before the PAUSED state, the hardware devices are not ready to doing a seek and reset the devices,
such as the video decoder, knowing that in the READY state we reserve and open the devices asynchronously.

We also need to avoid performing a new seek when another is pending, this behavior improves robustness when
performing many searches in a short period of time.

* Source/WebCore/platform/graphics/gstreamer/mse/MediaPlayerPrivateGStreamerMSE.cpp:
(WebCore::MediaPlayerPrivateGStreamerMSE::seek):
(WebCore::MediaPlayerPrivateGStreamerMSE::doSeek):
(WebCore::MediaPlayerPrivateGStreamerMSE::asyncStateChangeDone):
(WebCore::MediaPlayerPrivateGStreamerMSE::updateStates):