Crash under WebNotificationManager::didUpdateNotificationDecision() on the bots https://bugs.webkit.org/show_bug.cgi?id=245380
Reviewed by Ben Nham.

We recently started calling `testRunner.grantNotificationPermission(self.origin)` in
testharnessreport.js in 254620@main. This means that in some cases, we might now call
this function with a null/opaque origin.

This would cause crashes in WebNotificationManager::didUpdateNotificationDecision()
when trying to add the null origin to the HashMap. To address the crashes, we now
ignore null origins in WebNotificationManagerProxy::providerDidUpdateNotificationPolicy().

* Source/WebKit/UIProcess/Notifications/WebNotificationManagerProxy.cpp:
(WebKit::WebNotificationManagerProxy::providerDidUpdateNotificationPolicy):

Canonical link: https://commits.webkit.org/254651@main