[css-scroll-snap] focus should not take into account scroll margin for visibility check https://bugs.webkit.org/show_bug.cgi?id=245100 <rdar://99840839>
Reviewed by Simon Fraser.

When an element is focused, we perform a visibility calculation to determine if we need to scroll
the element into view. This visibility calculation should not take into account the scroll margin of
the element. This prevents a bug where we wont scroll an element into view that is offscreen because
the rect including its scroll margnin is within the viewport.

* LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-margin-visibility-check.html:
* Source/WebCore/page/FrameView.cpp:
(WebCore::FrameView::scrollToFocusedElementInternal):

Canonical link: https://commits.webkit.org/254732@main