[GTK] Layout tests with ANGLE backend require to use GLX to make Xvfb and software rasterization work https://bugs.webkit.org/show_bug.cgi?id=245288
Reviewed by NOBODY (OOPS!).

Gtk layout tests are using Xvfb and software rasterization, this is detected and uses the X11
backends and GLX in WebKit, we need to make sure we have the GLX backend compiled for ANGLE and
we use it in GTK for the GraphicsContextGLFallback class. That solves crashes in the Layout tests
for WebGL with ANGLE.

* Source/ThirdParty/ANGLE/CMakeLists.txt: Compile GLX backend when using
  gtk if x11 is enabled.
* Source/ThirdParty/ANGLE/PlatformGTK.cmake: Add compilation options for
  x11, use platform types in that case.
* Source/WebCore/platform/graphics/gbm/GraphicsContextGLFallback.cpp:
(WebCore::GraphicsContextGLFallback::platformInitializeContext): In case
of gtk and x11 enabled we configure the display to use the ANGLE GLX backend.