getComputedStyle() on transform property should return function list https://bugs.webkit.org/show_bug.cgi?id=23924
Reviewed by Antti Koivisto.

We should serialize the "transform" property to a series of individual functions to match
the individual operations found in the TransformOperations value. To determine when we should
be doing this, we pass a nullptr value as the renderer when calling valueForPropertyInStyle()
on the ComputedStyleExtractor in KeyframeEffect::getKeyframes() and WebAnimation::commitStyles().

Note the slight regression with first-letter and first-line tests where we now expose the "transform"
property not being rejected at parse-time, bug 245523 will address this.

* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/KeyframeEffect-getKeyframes.tentative-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/first-letter-allowed-properties-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/first-line-allowed-properties-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/transform-interpolation-computed-value-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/transform-interpolation-inline-value-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animation/commitStyles-expected.txt:
* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::getKeyframes):
* Source/WebCore/animation/WebAnimation.cpp:
(WebCore::WebAnimation::commitStyles):
* Source/WebCore/css/ComputedStyleExtractor.cpp:
(WebCore::computedTransform):

Canonical link: https://commits.webkit.org/254760@main