Implement last baseline for block elements in a flex container. https://bugs.webkit.org/show_bug.cgi?id=244734 rdar://99506835
Reviewed by NOBODY (OOPS!).

CSS Box Alignment Module Level 3 defines the last baseline of block
elements to be:  the dominant last baseline of the last in-flow line
box in the block container, or is taken from the last in-flow
block-level child in the block container that contributes a set of
last baselines, whichever comes first/last. The logic for this ends up
being very similar to the first baseline logic, but we are simply
starting from the last child and working our way up instead of the other
way around. Once we compute the last baseline value this way, we can
align the block elements in a flex container using this new value as
the ascent. We also need to take into consideration what the alignment
type is so that we can compute either the first baseline or last
baseline.

This patch also adds the “margin before” to the first baseline of table
elements since it was not being added before. This is needed in order to
properly align other elements alongside a table with margins. Without
this addition, we were rendering the expectation for both
css-flexbox/flexbox-align-self-baseline-horiz-001a and
css-flexbox/flexbox-align-self-baseline-horiz-001b incorrectly.

* LayoutTests/TestExpectations:
* Source/WebCore/rendering/RenderBlock.cpp:
(WebCore::RenderBlock::lastLineBaseline const):
* Source/WebCore/rendering/RenderBlock.h:
* Source/WebCore/rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::lastLineBaseline const):
* Source/WebCore/rendering/RenderBlockFlow.h:
* Source/WebCore/rendering/RenderBox.h:
(WebCore::RenderBox::lastLineBaseline const):
* Source/WebCore/rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::marginBoxAscentForChild):
(WebCore::alignmentOffset):
(WebCore::RenderFlexibleBox::layoutAndPlaceChildren):
* Source/WebCore/rendering/RenderTableSection.cpp:
(WebCore::RenderTableSection::firstLineBaseline const):